requirements.txt
requests[socks]
fastapi
uvicorn[standard]
beautifulsoup4
lxml
onionsearch
google-genai
python-dotenv
pydantic-settings
pydantic[email]